Update Variables.xml (#120)
Hi, I am new to Kodi and would like to help with improving the skin for Dutch people. I would like to propose to change '-' to ':' for Dutch rating images. I have tested with TinyMediaManager and this works and shows the Dutch rating images correctly.
I will also propose to add some Dutch language strings that are not present currently.
Thanks for the great skin!